// small-ads-frontend/src/app/types/BackendErrors.ts
export enum GeneralErrors {
  UNHANDLED_ERROR = 'UNHANDLED_ERROR'
}

export enum UserErrors {
  ERROR_WHILE_SEARCHING_USER_IN_DATABASE = 'ERROR_WHILE_SEARCHING_USER_IN_DATABASE'
}

export enum TaskErrors {
  ERROR_WHILE_SEARCHING_TASK_IN_DATABASE = 'ERROR_WHILE_SEARCHING_TASK_IN_DATABASE'
}

export enum AuthenticationErrors {
  AUTH_TOKEN_EXPIRED = 'AUTH_TOKEN_EXPIRED',
  AUTH_MUST_BE_LOGGED_OFF = 'AUTH_MUST_BE_LOGGED_OFF',
  AUTH_NO_TOKEN_PROVIDED = 'AUTH_NO_TOKEN_PROVIDED',
  AUTH_TOKEN_IS_NOT_AUTHENTIC = 'AUTH_TOKEN_IS_NOT_AUTHENTIC',
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_NO_ROLE_ALLOWED = 'AUTH_NO_ROLE_ALLOWED',
  AUTH_ACCESS_TO_INTRANET_NOT_ALLOWED = 'AUTH_ACCESS_TO_INTRANET_NOT_ALLOWED',
  AUTH_DISABLED_ACCOUNT = 'AUTH_DISABLED_ACCOUNT',
  ACCESS_NOT_AUTHORIZED = 'ACCESS_NOT_AUTHORIZED'
}